refactor(speech): split backends_common into mod.rs + tests.rs (#163)
backends_common.rs had grown to 554 lines — over the 500-line guideline —
solely because of its `#[cfg(test)] mod tests` block (the production
helpers are ~430 lines and read well together, as the file's own TODO
note observed). Move the `make_on_audio` unit tests into a sibling
`tests.rs` child module, following the `foo/{mod,tests}.rs` pattern, so
both files land under the guideline (mod.rs 435 / tests.rs 115) with no
production-code change.

- `git mv` preserves history (rename detected at 81% similarity).
- Test paths are unchanged (`voice_loop::backends_common::tests::*`).
- `pub(super)` production items keep the same visibility to sibling
  builder modules; only tests moved to a descendant module.

Verified from src/ with +1.88: fmt --all --check clean; clippy
-p primer-speech --features voice-loop and voice-loop,macos-native
--all-targets -D warnings clean; the 3 make_on_audio tests pass
identically to the pre-refactor baseline.

Closes #163.
